backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
audioread==3.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cloudinary==1.44.1
decorator==5.2.1
distro==1.9.0
fastapi==0.135.3
filelock==3.25.2
fsspec==2026.3.0
groq==1.1.2
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
joblib==1.5.3
lazy-loader==0.5
librosa==0.11.0
llvmlite==0.47.0
msgpack==1.1.2
numba==0.65.0
numpy==2.4.4
packaging==26.0
pillow==12.2.0
platformdirs==4.9.4
pooch==1.9.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
starlette==1.0.0
threadpoolctl==3.6.0
tokenizers==0.22.2
tqdm==4.67.3
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.43.0
